backend/requirements.txt
# FastAPI và ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database ORM và Migration
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic cho Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication và Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File và Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
face-recognition==1.3.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Caching và Session
redis==5.0.1
python-memcached==1.62

# Logging và Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment và Configuration
python-dotenv==1.0.0

# Background Tasks
celery==5.3.4
kombu==5.3.4

# Network và Discovery
zeroconf==0.131.0
netifaces==0.11.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Template Engine (nếu cần render HTML)
jinja2==3.1.2

# WebSocket support
websockets==12.0

# Rate Limiting
slowapi==0.1.9
